% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization_old.R
\name{create_plot_df}
\alias{create_plot_df}
\title{Create data frame for plotting}
\usage{
create_plot_df(data_list, stage, group_name = NULL)
}
\arguments{
\item{data_list}{List of methylation data}

\item{stage}{Label for normalization stage}

\item{group_name}{Name of the group (optional)}
}
\value{
Data frame with methylation rates and metadata
}
\description{
Create data frame for plotting
}
\keyword{internal}
