% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization_old.R
\name{create_plot_df_groups}
\alias{create_plot_df_groups}
\title{Create plot data frame for groups}
\usage{
create_plot_df_groups(data_list, stage, group_names)
}
\arguments{
\item{data_list}{List of methylation data}

\item{stage}{Label for normalization stage}

\item{group_names}{Vector of group names}
}
\value{
Data frame with methylation rates and group information
}
\description{
Create plot data frame for groups
}
\keyword{internal}
