% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creat_plots.R
\name{create_plots}
\alias{create_plots}
\title{Create Plots}
\usage{
create_plots(data, measure, ylabel, normalized = TRUE)
}
\arguments{
\item{data}{Prepared data for plotting.}

\item{measure}{Measurement type (e.g., "rate", "cov", "mc").}

\item{ylabel}{Label for the y-axis.}

\item{normalized}{Boolean indicating if the data is normalized.}
}
\value{
List of ggplot objects.
}
\description{
Create plots for methylation data.
}
