% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_within_groups.R
\name{.create_cor_plot}
\alias{.create_cor_plot}
\title{Internal function to create correlation heatmaps}
\usage{
.create_cor_plot(data, stage_name)
}
\description{
Internal function to create correlation heatmaps
}
\keyword{internal}
