% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{.get_optimal_layout}
\alias{.get_optimal_layout}
\title{Determine optimal layout for a set of plots}
\usage{
.get_optimal_layout(plot_types)
}
\arguments{
\item{plot_types}{Character vector of plot types}
}
\value{
A layout specification (ncol, nrow)
}
\description{
Determine optimal layout for a set of plots
}
\keyword{internal}
