% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_coverage.R
\name{.normalize_group}
\alias{.normalize_group}
\title{Internal function to normalize a group of samples}
\usage{
.normalize_group(replicates, sf)
}
\arguments{
\item{replicates}{List of replicate data frames}

\item{sf}{Vector of scaling factors}
}
\value{
List of normalized data frames
}
\description{
Internal function to normalize a group of samples
}
\keyword{internal}
