% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_utils.R
\name{.save_plot_report}
\alias{.save_plot_report}
\title{Save plot to PDF report}
\usage{
.save_plot_report(plot, plot_type, output_dir, n_facets)
}
\arguments{
\item{plot}{ggplot object}

\item{plot_type}{Plot type}

\item{output_dir}{Output directory}

\item{n_facets}{Number of facets (for layout)}
}
\value{
File path of saved report
}
\description{
Save plot to PDF report
}
\keyword{internal}
