% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_group_names.R
\name{extract_group_names}
\alias{extract_group_names}
\title{Extract group names from data}
\usage{
extract_group_names(data_list, diagnostics = FALSE)
}
\arguments{
\item{data_list}{A list of methylation data}

\item{diagnostics}{Whether to print diagnostic messages}
}
\value{
Character vector of group names
}
\description{
Extract group names from data
}
\keyword{internal}
