% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load and Preprocess Data}
\usage{
load_data(dir_path)
}
\arguments{
\item{dir_path}{Path to the directory containing the data files.}
}
\value{
A list containing GCH and HCG samples as data frames.
}
\description{
Load methylation data from files, preprocess, and split into GCH and
HCG samples. Files are expected to be named groupName_replicateName_site.tsv
Site is either GCH or HCG.
}
