% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_group.R
\name{normalize_group}
\alias{normalize_group}
\title{Normalize Group Data}
\usage{
normalize_group(replicates, sf)
}
\arguments{
\item{replicates}{List of replicate data frames.}

\item{sf}{Scaling factors to apply.}
}
\value{
List of normalized replicate data frames.
}
\description{
Normalize coverage and methylation counts for replicate groups.
}
