% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_methylation_rate.R
\name{normalize_methylation_within_set}
\alias{normalize_methylation_within_set}
\title{Internal function to normalize methylation rates within a set of samples}
\usage{
normalize_methylation_within_set(
  sample_list,
  alpha = 0.3,
  sites_per_quantile = 1000,
  max_quantiles = 50,
  diagnostics = TRUE
)
}
\arguments{
\item{sample_list}{List of replicate data frames}

\item{alpha}{Numeric between 0 and 1 controlling structure preservation (default = 0.3)}

\item{sites_per_quantile}{Target number of sites per quantile (default = 1000)}

\item{max_quantiles}{Maximum number of quantiles to use (default = 50)}

\item{diagnostics}{Logical indicating whether to print diagnostic information}
}
\description{
Internal function to normalize methylation rates within a set of samples
}
\keyword{internal}
