% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization.R
\name{plot_between_group_report}
\alias{plot_between_group_report}
\title{Create between-group normalization report}
\usage{
plot_between_group_report(
  within_norm,
  full_norm,
  group_names,
  output_file = NULL,
  color_scheme = "magma",
  bins = 50
)
}
\arguments{
\item{within_norm}{After within-group normalization}

\item{full_norm}{After both normalizations}

\item{group_names}{Vector of group names}

\item{output_file}{Optional file path to save report}

\item{color_scheme}{Color scheme for density plots}

\item{bins}{Number of hexagonal bins}
}
\value{
A grid arrangement of diagnostic plots
}
\description{
Create between-group normalization report
}
