% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization.R
\name{plot_boxplots}
\alias{plot_boxplots}
\title{Function to visualize distribution changes as Box plots}
\usage{
plot_boxplots(original_data, within_norm, full_norm)
}
\arguments{
\item{original_data}{Original data list before normalization}

\item{within_norm}{Data after within-group normalization}

\item{full_norm}{Data after both normalizations}
}
\value{
A ggplot object showing boxplots
}
\description{
Function to visualize distribution changes as Box plots
}
