% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization.R
\name{plot_rate_distributions}
\alias{plot_rate_distributions}
\title{Function to visualize density plots of rate distribution}
\usage{
plot_rate_distributions(original_data, within_norm, full_norm)
}
\arguments{
\item{original_data}{Original data list}

\item{within_norm}{Data after within-group normalization}

\item{full_norm}{Data after both normalizations}
}
\value{
A ggplot object showing density distributions
}
\description{
Function to visualize density plots of rate distribution
}
