% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization.R
\name{plot_within_group_report}
\alias{plot_within_group_report}
\title{Create within-group normalization report}
\usage{
plot_within_group_report(
  original_data,
  normalized_data,
  group_names,
  output_file = NULL,
  color_scheme = "magma",
  bins = 50
)
}
\arguments{
\item{original_data}{Original data list}

\item{normalized_data}{After within-group normalization}

\item{group_names}{Vector of group names}

\item{output_file}{Optional file path to save report}

\item{color_scheme}{Color scheme for density plots}

\item{bins}{Number of hexagonal bins}
}
\value{
A list containing both plot arrangements
}
\description{
Create within-group normalization report
}
