% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Visualization Functions for Methylation Data}
\description{
Functions for visualizing methylation data at various stages of the
normalization workflow. Includes density plots, QQ plots, MA plots,
and PCA/MDS plots.
}
