% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_plot_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare Data for Plotting}
\usage{
prepare_data(groups, measure)
}
\arguments{
\item{groups}{List of group replicate data frames.}

\item{measure}{The measurement to prepare data for (e.g., "rate", "cov", "mc").}
}
\value{
A data frame ready for plotting.
}
\description{
Prepare data frames for plotting normalization results.
}
