% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bedGraph.R
\name{process_single_file}
\alias{process_single_file}
\title{Process a single methylation file into bedGraph format}
\usage{
process_single_file(file_path, type, out)
}
\arguments{
\item{file_path}{Path to methylation file}

\item{type}{Type of bedGraph to create ("meth" or "cov")}

\item{out}{Output directory path}
}
\value{
Path to created bedGraph file
}
\description{
Process a single methylation file into bedGraph format
}
\keyword{internal}
