% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bedGraph.R
\name{process_single_sample}
\alias{process_single_sample}
\title{Process a single sample data frame into bedGraph format}
\usage{
process_single_sample(
  sample_df,
  type,
  out,
  group_name = NULL,
  sample_name = NULL
)
}
\arguments{
\item{sample_df}{Data frame containing methylation data}

\item{type}{Type of bedGraph to create ("meth" or "cov")}

\item{out}{Output directory path}

\item{sample_name}{Name of the sample (optional)}
}
\value{
Path to created bedGraph file
}
\description{
Process a single sample data frame into bedGraph format
}
\keyword{internal}
