% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_methylation_file}
\alias{read_methylation_file}
\title{Reads single methylation data file}
\usage{
read_methylation_file(file_path)
}
\arguments{
\item{file_path}{Path to methylation file}
}
\value{
data.table containing methylation data
}
\description{
Reads single methylation data file
}
\keyword{internal}
