% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_for_visualization.R
\name{sample_for_visualization}
\alias{sample_for_visualization}
\title{Sample methylation data for visualization}
\usage{
sample_for_visualization(
  original_data,
  within_norm,
  full_norm,
  n_sites = 1e+05,
  seed = 42
)
}
\arguments{
\item{original_data}{Original data list}

\item{within_norm}{Within-group normalized data}

\item{full_norm}{Fully normalized data}

\item{n_sites}{Number of sites to sample (default = 100000)}

\item{seed}{Random seed for reproducibility (default = 42)}
}
\value{
List containing sampled versions of input data maintaining the same structure
}
\description{
Sample methylation data for visualization
}
