% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_by_groups.R
\name{split_by_groups}
\alias{split_by_groups}
\title{Split data into Groups and Replicates}
\usage{
split_by_groups(data_list, group_names)
}
\arguments{
\item{data_list}{List containing GCH or HCG methylation data.}

\item{group_names}{Vector of group names to split by.}
}
\value{
A list containing GCH and HCG samples with replicates by group.
}
\description{
Splits data list into group specific lists so all replicates per group are together
}
