% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_shared_sites.R
\name{find_shared_sites}
\alias{find_shared_sites}
\title{Find Shared Sites}
\usage{
find_shared_sites(samples)
}
\arguments{
\item{samples}{List of sample data frames.}
}
\value{
A vector of shared site identifiers.
}
\description{
Identify shared sites across all samples within GCH or HCG data.
}
