% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization_old.R
\name{plot_ma}
\alias{plot_ma}
\title{Function to visualize differences vs average (MA plot)}
\usage{
plot_ma(original_data, within_norm, full_norm, max_points = 1e+05)
}
\arguments{
\item{original_data}{Original data list before normalization}

\item{within_norm}{Data after within-group normalization}

\item{full_norm}{Data after both normalizations}
}
\value{
A ggplot object showing MA plots
}
\description{
Function to visualize differences vs average (MA plot)
}
