% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_normalization.R
\name{plot_qq_comparison}
\alias{plot_qq_comparison}
\title{Function to visualize rates between samples by QQ plot}
\usage{
plot_qq_comparison(original_data, within_norm, full_norm)
}
\arguments{
\item{original_data}{Original data list before normalization}

\item{within_norm}{Data after within-group normalization}

\item{full_norm}{Data after both normalizations}
}
\value{
A ggplot object showing QQ comparisons
}
\description{
Function to visualize rates between samples by QQ plot
}
