% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bedGraph.R
\name{process_all_samples}
\alias{process_all_samples}
\title{Process all samples in a data list into bedGraph format}
\usage{
process_all_samples(data_list, type, out)
}
\arguments{
\item{data_list}{List of lists containing methylation data}

\item{type}{Type of bedGraph to create ("rate" or "cov")}

\item{out}{Output directory path}
}
\value{
Vector of paths to created bedGraph files
}
\description{
Process all samples in a data list into bedGraph format
}
\keyword{internal}
